+++ src/locale/es/_lib/formatLong/index.js
@@ -1,12 +1,41 @@
 import buildFormatLongFn from '../../../_lib/buildFormatLongFn/index.js'
 
-var formatLong = buildFormatLongFn({
-  LT: 'HH:mm',
-  LTS: 'HH:mm:ss',
-  L: 'DD/MM/YYYY',
-  LL: 'D [de] MMMM [de] YYYY',
-  LLL: 'D [de] MMMM [de] YYYY HH:mm',
-  LLLL: 'dddd, D [de] MMMM [de] YYYY HH:mm'
-})
+var dateFormats = {
+  full: "EEEE, d 'de' MMMM y",
+  long: "d 'de' MMMM y",
+  medium: 'd MMM y',
+  short: 'dd/MM/y'
+}
+
+var timeFormats = {
+  full: 'HH:mm:ss zzzz',
+  long: 'HH:mm:ss z',
+  medium: 'HH:mm:ss',
+  short: 'HH:mm'
+}
+
+var dateTimeFormats = {
+  full: "{{date}} 'a las' {{time}}",
+  long: "{{date}} 'a las' {{time}}",
+  medium: '{{date}}, {{time}}',
+  short: '{{date}}, {{time}}'
+}
+
+var formatLong = {
+  date: buildFormatLongFn({
+    formats: dateFormats,
+    defaultWidth: 'full'
+  }),
+
+  time: buildFormatLongFn({
+    formats: timeFormats,
+    defaultWidth: 'full'
+  }),
+
+  dateTime: buildFormatLongFn({
+    formats: dateTimeFormats,
+    defaultWidth: 'full'
+  })
+}
 
 export default formatLong